bug fixed

+++ EveryoneNews/Classes/Main(主要)/Model/Param/CardParam.m
@@ -18,7 +18,7 @@
 
 - (NSString *)startTime {
     if (_startTime == nil) {
-       return [NSString stringWithFormat:@"%ld", (long)([[NSDate date] timeIntervalSince1970] * 1000)];
+       return [NSString stringWithFormat:@"%lld", (long long)([[NSDate date] timeIntervalSince1970] * 1000)];
     }
     return _startTime;
 }
